requirements.txt
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
pytest-ordering
allure-pytest
PyYAML
selenium
requests
openpyxl
